separate loop_read_file to separate file
